Tests: Move script synchronization functions to utils library
