+++ tests/utils/utils.h
@@ -19,5 +19,7 @@
 #define TEST_UTILS_H
 
 int usleep_safe(useconds_t usec);
+int create_file(const char *path);
+int wait_on_file(const char *path);
 
 #endif /* TEST_UTILS_H */
